ai/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
transformers==4.38.1
torch==2.2.1
scikit-learn==1.4.1
pymysql==1.1.0
sqlalchemy==2.0.27
python-dotenv==1.0.1
nltk==3.8.1
spacy==3.7.2 